requirements.txt
PyGObject~=3.48.2
numpy~=1.23.5
matplotlib==3.8.4
PyQt5~=5.15.10
py-feat~=0.6.2
PySimpleGUI~=5.0.4
pillow==10.4.0
libreface~=0.1.0
scipy~=1.14.0
PyAutoGUI~=0.9.54